consul: Remove hardcoded http scheme
Consul seems to fill the config out with http scheme
by default. As of now libkv was also explicitely
putting the scheme to http thus conflicting with
the use of environment variables such as 'CONSUL_HTTP_SSL'.

This commit simply removes the line overriding the
http scheme thus using the default Consul behavior.